server/requirements.txt
uvicorn
fastapi[all]
pymongo
pyjwt
passlib[bcrypt]
selenium
redis
pymodm